requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
polars==0.18.8
pyjwt==1.7.1